+++ CHANGELOG.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
-- Fixed list style: Changes css selector `ul.list li` to `ul.list > li`.
+- Fixed list style: Changes css selector `ul.list li` to `ul.list > li`. #121
+- Fixed `.container` style. #120
 
 ## [0.6.1][] - 2018-04-24